requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.2
kombu==5.5.4
msgpack==1.1.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.7.1
scipy==1.16.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
xgboost==3.0.4
zope.interface==7.2
